review fixes: ISR safety, OTA guard, policy consts, const cast, FFI drop, fabric tests
radio_sx1262.c
  - Add portMUX_TYPE spinlock (s_dio1_mux) protecting s_dio1_pending.
    Atomically read-and-clear the flag in radio_service_rx() using
    portENTER_CRITICAL / portEXIT_CRITICAL so the dual-core race
    (ISR on CPU1 sets flag after check but before clear on CPU0) is
    eliminated.  ISR side uses portENTER_CRITICAL_ISR / portEXIT_CRITICAL_ISR.
  - Add #include "freertos/FreeRTOS.h" to make the dependency explicit.
  - Remove now-redundant s_dio1_pending = false inside fault-inject path.

rivr_ota_core.c
  - Add explicit underflow guard (payload_len < RIVR_OTA_SIG_LEN) directly
    before the size_t subtraction.  The earlier HDR_LEN check already makes
    this unreachable, but the local guard keeps the invariant visible and
    protects against future refactoring.

rivr_policy.c
  - Replace magic literals 1u/6u with PKT_CHAT/PKT_DATA symbolic constants
    in rivr_policy_allow_origination().

neighbor_table.c
  - Simplify const cast from (T*)(uintptr_t) round-trip to direct (T*) cast
    and improve comment: clarify that entries are always in writable BSS,
    never in read-only flash ROM.

rivr_core/src/ffi.rs
  - Drop the previous Engine before overwriting ENGINE_SLOT on re-init.
    Without this, a second rivr_engine_init() call before freeze leaked the
    scheduler, node vec, and binding map of the prior engine.

tests/
  - Add portMUX_TYPE / portENTER_CRITICAL* no-op stubs to host FreeRTOS stub
    so radio_sx1262.c compiles cleanly in test builds.
  - New test_fabric.c: 10 tests, 23 assertions covering every decision-
    threshold boundary in rivr_fabric_decide_relay():
      score <20 → SEND_NOW, =20 → DELAY(light,0ms), =49 → DELAY(light,290ms),
      =50 → DELAY(heavy,250ms), =79 → DELAY(heavy,540ms), =80 → DROP,
      =94 → DROP, =95 → DELAY(blackout guard,1000ms).
    Also tests control-packet guard rule and lifetime delta counters.
  - Makefile: add FABRIC_SRCS/BIN/FLAGS, fabric phony target, fabric binary
    rule, ASAN variant, and wire into all/clean.

+++ firmware_core/neighbor_table.c
@@ -203,10 +203,11 @@ const rivr_neighbor_t *neighbor_find(const rivr_neighbor_table_t *tbl,
         /* Expired — treat as absent. */
         if (age >= NTABLE_EXPIRY_MS) return NULL;
 
-        /* Update STALE flag as a derived/cached property.
-         * The cast-away of const is intentional: the flag is a pure
-         * function of existing data and changes no observable metric. */
-        rivr_neighbor_t *mutable_n = (rivr_neighbor_t *)(uintptr_t)n;
+        /* Update STALE flag as a cached/derived property.  Table entries
+         * live in writable BSS (never in read-only flash), so a direct
+         * non-const cast is well-defined here.  Using a direct cast rather
+         * than the (T*)(uintptr_t) round-trip to keep the intent clearer. */
+        rivr_neighbor_t *mutable_n = (rivr_neighbor_t *)n;
         if (age >= NTABLE_STALE_MS) {
             mutable_n->flags |= NTABLE_FLAG_STALE;
         } else {
